chngd

+++ database/migrations/2023_01_05_142904_create_contacts_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('email');
             $table->string('email_alt')->nullable();
             $table->text('address');
-            $table->unsignedInteger('person_id');
+            $table->unsignedBigInteger('person_id');
             $table->foreign('person_id')->references('id')->on('people');
             $table->timestamps();
         });
